Standardize GC env vars
